+++ src/components/date/__spec__.js
@@ -17,8 +17,8 @@ describe('Date', () => {
     );
   });
 
-  describe('intialize', () => {
-    it('sets the intial internal state', () => {
+  describe('initialize', () => {
+    it('sets the initial internal state', () => {
       expect(instance.state.open).toBeFalsy();
       expect(instance.state.datePickerValue).toBeNull();
       expect(instance.state.visibleValue).toEqual(today);
@@ -160,6 +160,17 @@ describe('Date', () => {
           expect(instance.setState).toHaveBeenCalledWith({ datePickerValue: '2015/01/01' });
         });
       });
+
+      describe('when date invalid', () => {
+        it('does not call setState', () => {
+          instance = TestUtils.renderIntoDocument(
+            <Date name='date' value='x' label='Date' />
+          );
+          spyOn(instance, 'setState');
+          instance.openDatePicker();
+          expect(instance.setState).not.toHaveBeenCalledWith({ datePickerValue: 'x' });
+        });
+      })
     });
   });