+++ src/components/date/date.js
@@ -15,6 +15,14 @@ import { validProps } from './../../utils/ether';
 import { tagComponent } from '../../utils/helpers/tags';
 
 /**
+ * Stores a reference to the current date in the given format,
+ * which can be used for default values.
+ *
+ * @property _today
+ * @type {string}
+ */
+const today = DateHelper.todayFormatted("YYYY-MM-DD");
+/**
  * A Date widget.
  *
  * == How to use a Date in a component:
@@ -66,11 +74,11 @@ class Date extends React.Component {
     /**
      * Sets the default value of the date field
      *
-     * @property defaultValue
+     * @property value
      * @type {String}
      * @default Today's date
      */
-    defaultValue: DateHelper.todayFormatted("YYYY-MM-DD"),
+    value: today,
 
     /**
     * Sets validations that should always be found on the component
@@ -106,9 +114,9 @@ class Date extends React.Component {
      *
      * @property visibleValue
      * @type {String}
-     * @default defaultValue
+     * @default value
      */
-    visibleValue: this.formatVisibleValue(this.props.value, this)
+    visibleValue: this.formatVisibleValue(this.props.value)
   }
 
   /**
@@ -128,14 +136,12 @@ class Date extends React.Component {
    * only when the field is not the active element.
    *
    * @method componentWillReceiveProps
-   * @param {Object} props The new props passed down to the component
+   * @param {Object} nextProps The new props passed down to the component
    * @return {void}
    */
-  componentWillReceiveProps(props) {
+  componentWillReceiveProps(nextProps) {
     if (this._document.activeElement != this._input) {
-      let value = props.value || props.defaultValue;
-      let date = this.formatVisibleValue(value, this);
-
+      let date = this.formatVisibleValue(nextProps.value);
       this.setState({ visibleValue: date });
     }
   }
@@ -214,7 +220,7 @@ class Date extends React.Component {
    * @return {void}
    */
   updateVisibleValue = () => {
-    let date = this.formatVisibleValue(this.props.value, this);
+    let date = this.formatVisibleValue(this.props.value);
     this.setState({
       visibleValue: date
     });
@@ -338,7 +344,6 @@ class Date extends React.Component {
     props.onKeyDown = this.handleKeyDown;
 
     delete props.autoFocus;
-    delete props.defaultValue;
     delete props.internalValidations;
 
     if (!this.props.readOnly && !this.props.disabled) {
@@ -359,15 +364,9 @@ class Date extends React.Component {
       ref: 'hidden',
       type: 'hidden',
       readOnly: true,
-      'data-element': 'hidden-input'
+      'data-element': 'hidden-input',
+      value: this.props.value
     };
-
-    if (typeof this.props.value !== 'undefined') {
-      props.value = this.props.value;
-    } else {
-      props.defaultValue = this.props.defaultValue;
-    }
-
 
     return props;
   }
@@ -506,28 +505,12 @@ class Date extends React.Component {
    *
    * @method formatVisibleValue
    * @param {String} value Unformatted Value
-   * @param {String} scope used to get default value of current scope if value doesn't exist
    * @return {String} formatted visible value
    */
-  formatVisibleValue(value, scope) {
-    value = value || this.getDefaultValue(scope);
+  formatVisibleValue(value) {
+    value = value || today;
     // Don't sanitize so it accepts the hidden format (with dash separators)
     return DateHelper.formatValue(value, this.visibleFormat(), { formats: this.hiddenFormat(), sanitize: false });
-  }
-
-  /**
-   * Returns defaultValue for specified scope,
-   *
-   * @method getDefaultValue
-   * @param {Object} scope used to get default value of current scope
-   * @return {String} default value
-   */
-  getDefaultValue(scope) {
-    if (typeof scope.refs.hidden !== 'undefined') {
-      return scope.refs.hidden.value;
-    } else {
-      return scope.props.defaultValue;
-    }
   }
 }
 ))));