+++ CHANGELOG.md
@@ -70,6 +70,7 @@ If you have Google Analytics enabled (`window.ga` is defined), and you are using
 * `Alert`: default size has been fixed to `extra-small`.
 * `ButtonToggle`: css typo corrected
 * `Confirm`: default size has been fixed to `extra-small`.
+* `Date`: fixed the warning about an uncontrolled input component
 * `Detail`: Footnote is allowed to expand vertically
 * `Heading`: alignment is fixed in IE where `hr` was centring by default
 * `Link`: CSS inheritance has been updated to better support buttons.